+++ src/VSIX/ApiClientCodeGen.Tests/Extensions/GetCustomToolNameTests.cs
@@ -1,19 +1,11 @@
-﻿using Rapicgen.Core;
-using Rapicgen.Extensions;
 using FluentAssertions;
+using Rapicgen.Core;
+using Rapicgen.Extensions;
 
 namespace Rapicgen.Tests.Extensions
 {
-    
     public class GetCustomToolNameTests
     {
-        [Xunit.Fact]
-        public void GetCustomToolName_AutoRest()
-            => SupportedCodeGenerator.AutoRest
-                .GetCustomToolName()
-                .Should()
-                .Contain("AutoRest");
-
         [Xunit.Fact]
         public void GetCustomToolName_NSwag()
             => SupportedCodeGenerator.NSwag