+++ src/VSIX/ApiClientCodeGen.Tests/NuGet/SupportedCodeGeneratorExtensionsTests.cs
@@ -1,10 +1,9 @@
-﻿using Rapicgen.Core;
-using Rapicgen.Core.Extensions;
 using FluentAssertions;
+using Rapicgen.Core;
+using Rapicgen.Core.Extensions;
 
 namespace Rapicgen.Tests.NuGet
 {
-    
     public class SupportedCodeGeneratorExtensionsTests
     {
         [Xunit.Fact]
@@ -17,13 +16,6 @@ namespace Rapicgen.Tests.NuGet
         [Xunit.Fact]
         public void GetDependencies_NSwagStudio_Returns_NotEmpty()
             => SupportedCodeGenerator.NSwagStudio
-                .GetDependencies()
-                .Should()
-                .NotBeNullOrEmpty();
-
-        [Xunit.Fact]
-        public void GetDependencies_AutoRest_Returns_NotEmpty()
-            => SupportedCodeGenerator.AutoRest
                 .GetDependencies()
                 .Should()
                 .NotBeNullOrEmpty();
@@ -57,13 +49,6 @@ namespace Rapicgen.Tests.NuGet
                 .Contain(c => c.Name == "Newtonsoft.Json");
 
         [Xunit.Fact]
-        public void GetDependencies_AutoRest_Contains_RestClientRuntime()
-            => SupportedCodeGenerator.AutoRest
-                .GetDependencies()
-                .Should()
-                .Contain(c => c.Name == "Microsoft.Rest.ClientRuntime");
-
-        [Xunit.Fact]
         public void GetDependencies_Swagger_Contains_RestSharp()
             => SupportedCodeGenerator.Swagger
                 .GetDependencies()
@@ -78,17 +63,17 @@ namespace Rapicgen.Tests.NuGet
                 .Contain(c => c.Name == "JsonSubTypes");
 
         [Xunit.Fact]
-        public void GetDependencies_OpenApi_Contains_RestSharp()
+        public void GetDependencies_OpenApi_Contains_MicrosoftExtensionsHosting()
             => SupportedCodeGenerator.OpenApi
                 .GetDependencies()
                 .Should()
-                .Contain(c => c.Name == "RestSharp");
+                .Contain(c => c.Name == "Microsoft.Extensions.Hosting");
 
         [Xunit.Fact]
-        public void GetDependencies_OpenApi_Contains_JsonSubTypes()
+        public void GetDependencies_OpenApi_Contains_MicrosoftExtensionsHttp()
             => SupportedCodeGenerator.OpenApi
                 .GetDependencies()
                 .Should()
-                .Contain(c => c.Name == "JsonSubTypes");
+                .Contain(c => c.Name == "Microsoft.Extensions.Http");
     }
 }