+++ src/VSIX/ApiClientCodegen.IntegrationTests/CustomTool/CSharpSingleFileCodeGeneratorYamlTests.cs
@@ -4,11 +4,9 @@ using ApiClientCodeGen.Tests.Common;
 using Rapicgen.Core;
 using Rapicgen.Core.Exceptions;
 using Rapicgen.Core.Options;
-using Rapicgen.Core.Options.AutoRest;
 using Rapicgen.Core.Options.General;
 using Rapicgen.Core.Options.OpenApiGenerator;
 using Rapicgen.Generators;
-using Rapicgen.Options.AutoRest;
 using Rapicgen.Options.General;
 using Rapicgen.Options.OpenApiGenerator;
 using FluentAssertions;
@@ -18,20 +16,6 @@ using Microsoft.VisualStudio.Shell.Interop;
 
 namespace Rapicgen.IntegrationTests.CustomTool
 {
-    [Trait("Category", "SkipWhenLiveUnitTesting")]
-    public class CSharpSingleFileCodeGeneratorYamlTests : TestWithResources
-    {
-        [SkippableFact(typeof(CustomToolException))]
-        public void AutoRest_CSharp_Test()
-        {
-            var optionsMock = new Mock<IAutoRestOptions>();
-            var optionsFactory = new Mock<IOptionsFactory>();
-            optionsFactory
-                .Setup(c => c.Create<IAutoRestOptions, AutoRestOptionsPage, DefaultAutoRestOptions>())
-                .Returns(optionsMock.Object);
-
-            Assert(SupportedCodeGenerator.AutoRest, optionsFactory.Object);
-        }
 
         [Fact]
         public void Swagger_CSharp_Test()