+++ src/VSIX/ApiClientCodegen.IntegrationTests/DependencyInstallerTests.cs
@@ -28,17 +28,6 @@ namespace Rapicgen.IntegrationTests
                 .Should()
                 .NotBeNullOrWhiteSpace();
 
-        [Fact]
-        public void InstallAutoRest_Returns_Path()
-        {
-            new Action(
-                    () => new DependencyInstaller(
-                        new NpmInstaller(new ProcessLauncher()),
-                        new FileDownloader(new WebDownloader()), new ProcessLauncher()).InstallAutoRest())
-                .Should()
-                .NotThrow();
-        }
-
         [SkippableFact(typeof(ProcessLaunchException))]
         public void InstallNSwag_Returns_Path()
         {