+++ src/VSMac/ApiClientCodeGen.VSMac/Commands/GenerateCode.cs
@@ -1,8 +1,7 @@
-﻿namespace ApiClientCodeGen.VSMac.Commands
+namespace ApiClientCodeGen.VSMac.Commands
 {
     public enum GenerateCode
     {
-        AutoRest,
         NSwag,
         NSwagStudio,
         Swagger,