+++ test/utilities.ps1
@@ -63,7 +63,7 @@ function Build-GeneratedCode {
     
     param (
         [Parameter(Mandatory=$true)]
-        [ValidateSet("All", "NSwag", "OpenApiGenerator", "SwaggerCodegen", "AutoRest-V2", "AutoRest-V3", "Kiota", "Refitter")]
+        [ValidateSet("All", "NSwag", "OpenApiGenerator", "SwaggerCodegen", "Kiota", "Refitter")]
         [string]
         $ToolName,
 
@@ -78,9 +78,9 @@ function Build-GeneratedCode {
     )
 
     if ($Version -eq "V2") {
-        $tools = @("AutoRest-V2", "NSwag", "Kiota", "Refitter", "SwaggerCodegen", "OpenApiGenerator")
+        $tools = @("NSwag", "Kiota", "Refitter", "SwaggerCodegen", "OpenApiGenerator")
     } else {
-        $tools = @("AutoRest-V3", "NSwag", "Kiota", "Refitter", "SwaggerCodegen", "OpenApiGenerator")
+        $tools = @("NSwag", "Kiota", "Refitter", "SwaggerCodegen", "OpenApiGenerator")
     }  
 
     if ($Parallel) {
@@ -97,9 +97,7 @@ function Build-GeneratedCode {
                 $argumentsList += "build ./GeneratedCode/$_/Net481/Net481.csproj"
                 $argumentsList += "build ./GeneratedCode/$_/Net472/Net472.csproj"
 
-                if ($_ -notcontains "AutoRest-V3") {
-                    $argumentsList += "build ./GeneratedCode/$_/Net462/Net462.csproj"
-                }
+                $argumentsList += "build ./GeneratedCode/$_/Net462/Net462.csproj"
             }
         } else {
             $argumentsList = @(
@@ -114,9 +112,7 @@ function Build-GeneratedCode {
                 "build ./GeneratedCode/$ToolName/Net472/Net472.csproj"
             )
 
-            if ($ToolName -notcontains "AutoRest-V3") {
-                $argumentsList += "build ./GeneratedCode/$ToolName/Net462/Net462.csproj"
-            }
+            $argumentsList += "build ./GeneratedCode/$ToolName/Net462/Net462.csproj"
         }
         
         $processes = ($argumentsList | ForEach-Object {
@@ -142,10 +138,7 @@ function Build-GeneratedCode {
                 dotnet build ./GeneratedCode/$_/Net481/Net481.csproj; ThrowOnNativeFailure
                 dotnet build ./GeneratedCode/$_/Net472/Net472.csproj; ThrowOnNativeFailure
 
-                if ($_ -notcontains "AutoRest-V3") {
-                    dotnet build ./GeneratedCode/$_/Net462/Net462.csproj; ThrowOnNativeFailure
-                    dotnet build ./GeneratedCode/$_/Net452/Net452.csproj; ThrowOnNativeFailure
-                }
+                dotnet build ./GeneratedCode/$_/Net462/Net462.csproj; ThrowOnNativeFailure
             }
         } else {
             Write-Host "`r`nBuilding $ToolName`r`n"
@@ -159,10 +152,7 @@ function Build-GeneratedCode {
             dotnet build ./GeneratedCode/$ToolName/Net481/Net481.csproj; ThrowOnNativeFailure
             dotnet build ./GeneratedCode/$ToolName/Net472/Net472.csproj; ThrowOnNativeFailure
 
-            if ($ToolName -notcontains "AutoRest-V3") {
-                dotnet build ./GeneratedCode/$ToolName/Net462/Net462.csproj; ThrowOnNativeFailure
-                dotnet build ./GeneratedCode/$ToolName/Net452/Net452.csproj; ThrowOnNativeFailure
-            }
+            dotnet build ./GeneratedCode/$ToolName/Net462/Net462.csproj; ThrowOnNativeFailure
         }
     }
 }
@@ -171,7 +161,7 @@ function Generate-Code {
     
     param (
         [Parameter(Mandatory=$true)]
-        [ValidateSet("NSwag", "OpenApiGenerator", "SwaggerCodegen", "AutoRest-V2", "AutoRest-V3", "Kiota", "Refitter")]
+        [ValidateSet("NSwag", "OpenApiGenerator", "SwaggerCodegen", "Kiota", "Refitter")]
         [string]
         $ToolName,
 
@@ -198,12 +188,6 @@ function Generate-Code {
         "OpenApiGenerator" { 
             $command = "csharp openapi -v V7140"
         }
-        "AutoRest-V2" {
-            $command = "csharp autorest"
-        } 
-        "AutoRest-V3" {
-            $command = "csharp autorest"
-        }
         "NSwag" {
             $command = "csharp nswag"
         }
@@ -213,10 +197,6 @@ function Generate-Code {
         "Refitter" {
             $command = "csharp refitter"
         }
-    }
-
-    if ($Version -eq "V3" -and $ToolName -eq "AutoRest-V2") {
-        $ToolName = "AutoRest-V3"
     }
 
     $project = "../src/CLI/ApiClientCodeGen.CLI/ApiClientCodeGen.CLI.csproj"
@@ -267,7 +247,7 @@ function Generate-CodeParallel {
     )
 
     $processes = @()
-    "AutoRest-V2", "NSwag", "SwaggerCodegen", "OpenApiGenerator", "Kiota", "Refitter" | ForEach-Object {
+    "NSwag", "SwaggerCodegen", "OpenApiGenerator", "Kiota", "Refitter" | ForEach-Object {
         switch ($_) {
             "SwaggerCodegen" {
                 $command = "csharp swagger"
@@ -299,7 +279,7 @@ function Generate-CodeParallel {
         $process.WaitForExit()
     }
 
-    "AutoRest-V2", "NSwag", "SwaggerCodegen", "OpenApiGenerator", "Kiota", "Refitter" | ForEach-Object {
+    "NSwag", "SwaggerCodegen", "OpenApiGenerator", "Kiota", "Refitter" | ForEach-Object {
         if (Test-Path "GeneratedCode/$_/Output.cs" -PathType Leaf) {
             Copy-Item "GeneratedCode/$_/Output.cs" "./GeneratedCode/$_/Net7/Output.cs" -Force
             Copy-Item "GeneratedCode/$_/Output.cs" "./GeneratedCode/$_/Net6/Output.cs" -Force
@@ -322,7 +302,7 @@ function Generate-CodeThenBuild {
     
     param (
         [Parameter(Mandatory=$false)]
-        [ValidateSet("All", "NSwag", "OpenApiGenerator", "SwaggerCodegen", "AutoRest-V2", "AutoRest-V3", "Kiota", "Refitter")]
+        [ValidateSet("All", "NSwag", "OpenApiGenerator", "SwaggerCodegen", "Kiota", "Refitter")]
         [string]
         $ToolName = "All",
 
@@ -352,9 +332,9 @@ function Generate-CodeThenBuild {
             Build-GeneratedCode -ToolName $ToolName
         } else {
             if ($Version -eq "V2") {
-                $tools = @("NSwag", "Kiota", "Refitter", "OpenApiGenerator", "SwaggerCodegen", "AutoRest-V2")
+                $tools = @("NSwag", "Kiota", "Refitter", "OpenApiGenerator", "SwaggerCodegen")
             } else {
-                $tools = @("NSwag", "Kiota", "Refitter", "OpenApiGenerator", "SwaggerCodegen", "AutoRest-V3")
+                $tools = @("NSwag", "Kiota", "Refitter", "OpenApiGenerator", "SwaggerCodegen")
             }  
             $tools | ForEach-Object {
                 Generate-CodeThenBuild `